requirements.txt
beautifulsoup4==4.9.3
IMAPClient==2.1.0
Pillow==8.0.1
PyPDF2==1.26.0
pytest==6.1.1
pyzmail36==1.0.4
requests==2.24.0
selenium==3.141.0